+++ .eslintrc
@@ -98,6 +98,7 @@
     "no-catch-shadow": "error",
     "no-delete-var": "error",
     "no-label-var": "error",
+    "no-restricted-globals": "off",
     "no-shadow-restricted-names": "error",
     "no-shadow": "off",
     "no-undef-init": "error",
