+++ src/shared/streams_polyfill.js
@@ -12,6 +12,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
+/* eslint-disable no-restricted-globals */
 
 let isReadableStreamSupported = false;
 if (typeof ReadableStream !== 'undefined') {
