+++ web/firefoxcom.js
@@ -14,7 +14,7 @@
  */
 
 import '../extensions/firefox/tools/l10n';
-import { createObjectURL, PDFDataRangeTransport, shadow } from 'pdfjs-lib';
+import { createObjectURL, PDFDataRangeTransport, shadow, URL } from 'pdfjs-lib';
 import { BasePreferences } from './preferences';
 import { PDFViewerApplication } from './app';
